requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-cors-headers==4.4.0
gunicorn==21.2.0
